[scripts] always commit after a change
This is done so that we have the entire history.
Also we do have a separate marker for manual commits but since we do not
have anything to commit, not sure how this will be usefule. This is
mostly useful in case of tk, but I am thinking about calling note after
running tk.

In case this ever gets really big to manage, we can just always squash
the commits.