[scripts] hacky version of floating tmuxproject
We could not get the exact sizes as if we tried to get them smaller tmux
does not properly start the splits while in daemon mode thinking there
is not space. If we change from daemon mode, then nothing works.

As of now, we can get the layout more or less, mainly just have to
adjust the height of the middle window.