[scripts] add onacall
We are technically only checking if zoom is running, not necessarily in
a call. This will prevent it returning true even if not in a call.

Initially I thought this would be a bad idea, but then I only leave zoom
open if I am expecting most of the day to be a call and I would not have
time to deal with `randomidea` anyways as that is what I use this script
for anyways. So, not a bug just a feature I guess.