[scripts] note: make temp listing show title for grepping
`temp` notes do not have a good filename, so we kinda need the title to do
a proper search on it when using fzf. I thought of adding this in for
all the note listing but that looked really bad. `temp` note filenames
are all the same length and so this actually look good.

Maybe we could enable it for quote and other generated filenames, got
time to think about them though.