[scripts] taskwatch: disable --force-poll
This was added thinking it would do a refresh after that amount of time,
but turns out this is just a force of the check if anything changed.
As of now to mitigate this, I have added in a cronjob to run every 5
minutes which should refresh the db and unwait recuring jobs.