[hammerspoon] focusandback: move mouse to the resulting screen
Maybe we could move mouse to the center of the app that is getting
focus, but let me try this out first and see how this goes for now.

We are having extra checks to see if we are on the same scren as
before as we don't want to reset the mouse position in case we don't
change screens.