[emacs] eglot: enable eldoc now that it is limited to one line
Previously I had disabled eldoc mode because it was causing a lot of
change in the window arrangement if the doc was more than one line,
now we have change eldoc to only use single line and this has become
useable (and even useful). So enabling it. Related commit: ecf7e8b