[scripts] add task-to-html
This will let us convert task items to html. Used as a cron job which
will output this to a specified directory. From there we sync this
using syncthing to the phone and use and app which is supposed to
display html pages as widgets to display this in the home screen.