[scripts] markdown-preview: initial commit
A tool that can be used to preview markdown as github flavoured html
document. It is not an exact replica of github, but the core idea
needed for transformation, refresh and all that jazz is there. All we
now need to make it exactly github like is just proper css.