[emacs] remove grip for markdown preview
Grip is actually calling out to github apis to render this, not
something that I want. Plus now that we have the markdown-preview
script we are much better off anyways using an offline only solution.
With the proper css the markdown-preview script might be better than
grip. Looking forward to that.