[nvim] making `d` yank to clipboard
This is not what I am used to in the vim world, but since moving to
Emacs and evil mode, I tried this out and this seems to be better
actually. So sticking with this in vim.

One difference between this and emacs setup is that in emacs current
even `c` and `x` go to clipboard. I am still just deciding if that
should be ported over to here as well. But as of now, `d` is necessary.