[emacs] eglot: fix issue with project.el
Previously eglot was throwing the below error.
```
eglot--connect: Symbol’s function definition is void: project-root
```

This was because emacs was loading project.el that was bundled with
emacs whcih for some reason did not have that function. We had to use
project.el from melpa, but since the load with straight was much lower
some package was requiring this package and it was loading the default
emacs version. Now that the straight load of this package was moved
up, it actually loads the proper version of the package.