[emacs] disabling menu-bar-mode is messy with server
I can check for window system, but since I start emacs using server
the window sysetem if false even even I indent to use a gui frame. I
could probably control this per frame, but that is for a future me.

I primary wanted to do this so that I can use it in the floating
terminal, but for some reason Emacs gets completely stuck and I did't
really wanna bother with it as of now.