[emacs] properly disable tramp
I really wish tramp would work, but most of the time it just freezes
and even if it does not freeze it is too slow to be usable.

I am not sure if this is becase the other plugins are interacting with
it and messing this up. I just have not been able to get this thigns
working reliability yet.