[emacs] tramp: temporarily re enable tramp stuff
There are some additional fixes in other places to hopefully speed up
tramp. It is still a bit buggy, but now we can use it at times.

What I generally found was that callilng projectiles messes things up
when we are dealing with tramp. The fixes are mostly to avoid doing
anything with projectils when using tramp.

Btw use `pkill -SIGUSR2 Emacs` to get out when tramp hangs emacs.