[nix] use dragon to simplify sharing files, primarily screenshot
Dragon lets us create a drag and drop source, wired this up with tmux
and mouse clicks. Now when we want to share a file we have in a
terminal in the current directory, we just have to double click on
that and we can share it using drag and drop.

There is also now a keybinding to open up the last screenshot that we
took in dragon for sharing that easily.