[scripts] ,tmux-popup-opener: having it send commands is a bad idea
If we are in a chat application like gomuks and I for some reason
double click on something, this will end up sending a message on
whatever channel that I am in. Plus if I going to do this I might as
well just open a graphical file manager. Also, I never got the going
back thing working. In short it was not all that useful but was risky,
so dropping it in favour of using dragon for sharing folder.