[i3] set a wallpaper on startup using feh
When landing on a new session, since it is just a black screen with
nothing in the way to tell me that I have actually booted in, no
wallpaper, no bar, no notifications etc it might be confusing. This at
least gives some visual indication that I have booted up properly.