[emacs] make golangci-lint work with flymake
It kinda gets annoying as the time it takes to rerun the checks are
long and most of the time the buffer has incorrect stuff, but at least
now it is working. I might end up disabling it if it starts being
annoying.

The major difference here is that golangci-lint needs a project
context and not just the file, so we are kinda hacking that in as that
is not the expected behavior with the flymake-quickdef package.