[emacs] add a way to do quick calculations on region
You can now just select a region and hit `s-:` to get the output. If
not region is selected, then it uses the current line. If you prefix
with a `C-u`, we output the result to the buffer. Not sure if we
should replace the buffer instead of just appending it, but I will
think about that later.