[emacs] go projects can now be recognized via go.mod
Previously when using lsp and we go to a vendored package without git
tracking, it would not be able to do any kind of project operations,
main one being `project-find-files` and lsp. This should now be fixed
now that we are able to detect project root for go projects by looking
for go.mod file in parent directories.