[emacs] add GH copilot config
Well, just trying it out. Keeping the timeout to 2s so that it does
not get in the way, but if I stop to think,it will suggest me some
options. From briefly using it for a few days, it looks like it is
good for repetitive things that you don't wanna do or giving you ideas
on how to proceed. Using it for writing actually new code only happen
to create very subtle bugs that was harder to debug and fix.