[emacs] drop elfeed
I was running into a lot of cases where the elfeed db would not
load. Finally got tired of it and was not in the mood to debug
it. There is an open ticket which is not getting any attention. Also,
the last commit to elfeed seems to be quite a while ago.

elfeed-db-load: Wrong type argument: avl-tree-, nil
Similar to https://github.com/skeeto/elfeed/issues/340

Ended up self hosting miniflux, which also has the advantage of being
available on my phone.