[scripts] add script to create a git worktree in tmp
This, given a commit starts a creates a git worktree on the commit in
tmp. Using commit so that I won't get clashes at a later point if I
try to switch to that branch. This is mostly just used of opening a
different version of the project as a separate instance to view how
things where.