[emacs] split up certain parts of the config
This is part of a longer term effort to split up my gigantic emacs
init file. I am less than half way through. The config file was
getting pretty gnarly to manage.  Let's split it up into multiple
files so that it is easier to manage and read (for me and llms).